refactor(Lotto): 로또 클래스 패키지 변경
